requirements.txt
--find-links wheels
git+https://github.com/meraki-analytics/cassiopeia
cassiopeia-diskstore
pycurl
certifi
tqdm
